chore(makefile): add simple makefile
The makefile uses the ctags program from brew with the recursive option
to generate a tags file for the whole project. I have instructed the stderror
to be sent to /dev/null, because this program is not installed by
default so I don't want it to print garbage if this doesn't work.