feat(launch_function): launch and save return
the launch_function function will fork the current process, call the function
pointer with the test case, and then wait for the child to finish to
access it's value. It does not yet check if the process seg faulted or
bus errored or anything, and at the moment I'm not actually convinced
that it properly gets the return status either. But this commit also
enables the launch_tests function to call multiple tests, and free each
node and the final queue when it's completed.