feat(launch_tests): keep track of pass vs fail
Change a bunch of things to keep track of whether or not a test succeeded
so we can total that up as we go. Also is needed so we can check the return
value to see if the whole test group passed or not. Also some norming,
I had to move the signal handling bit to it's own function because it
wouldn't fit in with the rest of it.